chore: reorganize test files into new tests directory
