+++ tests/unit/components/HelloWorld.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import HelloWorld from './HelloWorld';
+import HelloWorld from '../../../src/components/HelloWorld';
 
 describe('HelloWorld component', () => {
   it('renders the correct text', () => {
